+++ opentariff/Enums/base_enums.py
@@ -1,0 +1,16 @@
+class EnumBase:
+    """Base class for enums to ensure consistent string representation"""
+
+    @classmethod
+    def values(cls) -> list:
+        return [member.value for member in cls]
+
+
+class DayOfWeek(int, EnumBase):
+    MONDAY = 0
+    TUESDAY = 1
+    WEDNESDAY = 2
+    THURSDAY = 3
+    FRIDAY = 4
+    SATURDAY = 5
+    SUNDAY = 6
